move parquet constructor details into ResponseSink::Parquet constructor
